added 2 sample files dealing with auditing executions and errors
these are some classes that are part of a larger project.  ErrorActionItem.cs and ErrorManager.cs deal with logging errors to the database and displaying to user. RunAuditActionItem.cs is responsible for logging audits of user-initiated executions against the database, including username, time, what was executed, against what, and whether successful.